feat(dynamic-sampling): Add error rule modal
